+++ packages/react/src/components/DateTimePicker/DateTimePickerV2WithoutTimeSpinner.jsx
@@ -30,7 +30,7 @@ import {
 } from '../../constants/DateConstants';
 import Button from '../Button/Button';
 import FlyoutMenu, { FlyoutMenuDirection } from '../FlyoutMenu/FlyoutMenu';
-import { handleSpecificKeyDown } from '../../utils/componentUtilityFunctions';
+import { handleSpecificKeyDown, useOnClickOutside } from '../../utils/componentUtilityFunctions';
 import { Tooltip } from '../Tooltip';
 
 import {
@@ -46,6 +46,8 @@ import {
   useDateTimePickerRef,
   useDateTimePickerTooltip,
   useRelativeDateTimeValue,
+  useDateTimePickerClickOutside,
+  useCloseDropdown,
 } from './dateTimePickerUtils';
 
 const { iotPrefix, prefix } = settings;
@@ -328,6 +330,7 @@ const DateTimePicker = ({
   const [focusOnFirstField, setFocusOnFirstField] = useDateTimePickerFocus(datePickerElem);
   const relativeSelect = useRef(null);
   const containerRef = useRef();
+  const dropdownRef = useRef();
   const updatedStyle = useMemo(
     () => ({
       ...style,
@@ -372,7 +375,7 @@ const DateTimePicker = ({
     onNavigateRadioButton,
     onNavigatePresets,
   } = useDateTimePickerKeyboardInteraction({ expanded, setCustomRangeKind });
-  const [isTooltipOpen, toggleTooltip] = useDateTimePickerTooltip({ isExpanded });
+  const [isTooltipOpen, toggleTooltip, setIsTooltipOpen] = useDateTimePickerTooltip({ isExpanded });
 
   const dateTimePickerBaseValue = {
     kind: '',
@@ -652,6 +655,28 @@ const DateTimePicker = ({
     }
   };
 
+  // Close tooltip if dropdown was closed by click outside
+  const onFieldBlur = (evt) => {
+    if (evt.target !== evt.currentTarget) {
+      setIsTooltipOpen(false);
+    }
+  };
+
+  const closeDropdown = useCloseDropdown({
+    isExpanded,
+    isCustomRange,
+    setIsCustomRange,
+    setIsExpanded,
+    parseDefaultValue,
+    defaultValue,
+    setCustomRangeKind,
+    lastAppliedValue,
+  });
+
+  const onClickOutside = useDateTimePickerClickOutside(closeDropdown, containerRef);
+
+  useOnClickOutside(dropdownRef, onClickOutside);
+
   // eslint-disable-next-line react/prop-types
   const CustomFooter = () => {
     return (
@@ -755,7 +780,7 @@ const DateTimePicker = ({
           onFieldInteraction(event);
         })}
         onFocus={toggleTooltip}
-        onBlur={toggleTooltip}
+        onBlur={onFieldBlur}
         onMouseEnter={toggleTooltip}
         onMouseLeave={toggleTooltip}
         tabIndex={0}
@@ -846,6 +871,7 @@ const DateTimePicker = ({
             style={updatedStyle}
           >
             <div
+              ref={dropdownRef}
               className={`${iotPrefix}--date-time-picker__menu-scroll`}
               style={{ '--wrapper-width': '20rem' }}
               role="listbox"